splitting controllers
