+++ src/main/java/FC/Controller.java
@@ -10,43 +10,17 @@ public class Controller {
     private final String menuTable = "menu";
     private final String dBFile = "flash_cards.db";
     private final Connection connection = new Connector(dBFile).connect();
-    private Deck unMastered;
-    private Deck mastered;
-    private final Random rand = new Random();
-
-    /**
-     * Returns the next card to study
-     */
-    public Card getNextToStudy() {
-        return unMastered.getCard(rand.nextInt(unMastered.getSize()) - 1);
-    }
-
-    /**
-     * Moves card from unMastered to mastered
-     */
-    public void masterCard(Card card) {
-        unMastered.removeCard(card);
-        mastered.addCard(card);
-    }
-
-    /**
-     * Loads new deck to unMastered and clears mastered
-     */
-    public void startNewStudySession(String deckName) throws SQLException{
-        unMastered = getDeck(deckName);
-        mastered.clear();
-    }
 
     /**
      * Create a new table/deck in the database
      */
-    private void addDeck(String title) throws SQLException {
+    protected void addDeck(String title) throws SQLException {
         String createTableStmt = "CREATE TABLE " + title + " (term TEXT NOT NULL, def TEXT NOT NULL);";
         executeCUD(createTableStmt);
         addDeckToMenuTable(title);
     }
 
-    private void addDeckToMenuTable(String title) throws SQLException {
+    protected void addDeckToMenuTable(String title) throws SQLException {
         String insertToMenuStmt = "INSERT INTO " + menuTable + "VALUES (" + title + ");";
         executeCUD(insertToMenuStmt);
     }
@@ -54,7 +28,7 @@ public class Controller {
     /**
      * Delete a card/row from a given deck/table
      */
-    private void deleteCard(Card card, String table) throws SQLException {
+    protected void deleteCard(Card card, String table) throws SQLException {
         String removeCardStmt = "DELETE FROM " + table + " WHERE rowid = " + card.getId() + ";";
         executeCUD(removeCardStmt);
     }
@@ -62,7 +36,7 @@ public class Controller {
     /**
      * Update a given card's term
      */
-    private void updateTerm(String table, Card card) throws SQLException {
+    protected void updateTerm(String table, Card card) throws SQLException {
         String updateTermStmt = "UPDATE " + table + "SET term = " + card.getTerm() + "WHERE rowid = " + card.getId() + ";";
         executeCUD(updateTermStmt);
     }
@@ -70,7 +44,7 @@ public class Controller {
     /**
      * Update a given card's def
      */
-    private void updateDef(String table, Card card) throws SQLException {
+    protected void updateDef(String table, Card card) throws SQLException {
         String updateDefStmt = "UPDATE " + table + "SET def = " + card.getDef() + "WHERE rowid = " + card.getId() + ";";
         executeCUD(updateDefStmt);
     }
@@ -78,13 +52,13 @@ public class Controller {
     /**
      * Delete a table/deck in the database
      */
-    private void deleteDeck(String table) throws SQLException{
+    protected void deleteDeck(String table) throws SQLException{
             String deleteDeckStmt = "DROP TABLE " + table + ";";
             executeCUD(deleteDeckStmt);
             removeFromMenu(table);
     }
 
-    private void removeFromMenu(String table) throws SQLException {
+    protected void removeFromMenu(String table) throws SQLException {
         String deleteMenuItemStmt = "DELETE FROM " + menuTable + " WHERE deck_title = " + table + ";";
         executeCUD(deleteMenuItemStmt);
     }
@@ -92,19 +66,19 @@ public class Controller {
     /**
      * Add card to an existing table/deck
      */
-    private void insertCard(String table, String term, String def) throws SQLException{
+    protected void insertCard(String table, String term, String def) throws SQLException{
             String insertCardStmt = "INSERT INTO " + table + "VALUES (" + term + ", " + def + ");";
             executeCUD(insertCardStmt);
     }
 
-    private void executeCUD(String cudStatement) throws SQLException {
+    protected void executeCUD(String cudStatement) throws SQLException {
         connection.createStatement().execute(cudStatement);
     }
 
     /**
      * @return Array of all decks in the db
      */
-    private String[] getAllDecks() throws SQLException{
+    protected String[] getAllDecks() throws SQLException{
         List<String> deckList = new ArrayList<>();
         ResultSet results = selectAll(menuTable);
 
@@ -115,7 +89,7 @@ public class Controller {
         return (String[]) deckList.toArray();
     }
 
-    private ResultSet selectAll(String table) throws SQLException {
+    protected ResultSet selectAll(String table) throws SQLException {
         String selectStmt = "SELECT * FROM " + table;
         return connection.createStatement().executeQuery(selectStmt);
     }
@@ -123,13 +97,13 @@ public class Controller {
     /**
      * populate a deck object using given deck's table
      */
-    private Deck getDeck(String table) throws SQLException{
+    protected Deck getDeck(String table) throws SQLException{
         Deck deck = new Deck();
         getDeckData(table, deck);
         return deck;
     }
 
-    private void getDeckData(String table, Deck deck) throws SQLException{
+    protected void getDeckData(String table, Deck deck) throws SQLException{
         ResultSet results = selectAll(table);
 
         while (results.next()) {
@@ -137,7 +111,7 @@ public class Controller {
             }
     }
 
-    private void addCardFromDB(Deck deck, ResultSet rs) throws SQLException {
+    protected void addCardFromDB(Deck deck, ResultSet rs) throws SQLException {
         deck.addCard(new Card(rs.getString("rowid"),
                 rs.getString("term"), rs.getString("def")));
     }
