+++ src/main/java/FC/EditController.java
@@ -1,0 +1,10 @@
+package FC;
+
+public class EditController extends Controller{
+
+
+
+    public EditController(){
+
+    }
+}
